Update App.js for issue #3 Number Guessing Game : Allow players to request hints (e.g., "The number is even") for the number guessing game
Description:
This PR refers to issue #3 Number Guessing Game : Allow players to request hints (e.g., "The number is even") for the number guessing game 

Modifications Done:
Modified the code and provided a simple user-friendly interface to the game. 
Also added a button for the user to request for the hint. The hint is provided max for 3 times then the game gets over and the original number is revealed.