Update App.css for issue GDG-IGDTUW#3 Number Guessing Game : Allow players to request hints (e.g., "The number is even") for the number guessing game
Update App.css for issue GDG-IGDTUW#3 Number Guessing Game : Allow players to request hints (e.g., "The number is even") for the number guessing game
Description:
This PR refers to issue GDG-IGDTUW#3 Number Guessing Game : Allow players to request hints (e.g., "The number is even") for the number guessing game 

Modifications Done:
The css file is updated to provide styling to the react app. 
This helps in better visibility and interaction with the game.